backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
pandas==2.2.1
openpyxl==3.1.2
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
python-dotenv==1.0.1 